requirements.txt
# Add your project dependencies here

# Requirements for vibe-sandbox
streamlit
pyyaml
jsonschema
streamlit-tree-select
